+++ game/cli.py
@@ -42,16 +42,20 @@ class Game():
                             scrabbleGame.current_player.fillTiles(scrabbleGame.bagTiles)
                             break;
                         elif choice == "2":
-                            print("El formato para elegir es del 1 al 7, pudiendo elegir un rango de ellas o una sola")
-                            tiles = input("Elija qué tiles cambiar!")
-                            ## función de cambiar tiles
-                            ## volver al loop inicial, cómo?
+                            list = input("Elija qué tiles cambiar! Separe con coma, por ej: 1, 5, 3: ")
+                            res = [eval(i) for i in list.split(', ')]
+                            scrabbleGame.current_player.swapTiles(scrabbleGame.bagTiles, res)
+                            print(scrabbleGame.current_player.tiles)
+                            print("Tiles cambiadas!")
+                            time.sleep(1.5)
                         elif choice == "3":
                             scrabbleGame.current_player.shuffleTiles()
                             print(scrabbleGame.current_player.tiles)
                             print("Shuffled!")
                             time.sleep(1.5)
                         elif choice == "4":
+                            print("Pasaste tu turno!")
+                            time.sleep(1.5)
                             break;
                         elif choice == "5":
                             exit = input("Está seguro de terminar la partida? Y/N: ")
@@ -72,7 +76,7 @@ class Game():
                 print("Valor equivocado!")
         else:
             input("Juego terminado! Desea jugar otra vez? (Y/N): ")
-        
+
     def showBoard(self, board):
         print('\n  |' + ''.join([f' {str(row_index).rjust(2)} ' for row_index in range(1, 16)]))
         for row_index, row in enumerate(board.grid):